+++ lib/federation_kit.rb
@@ -18,6 +18,7 @@ module FederationKit
 
     include plugin::InstanceMethods if defined?(plugin::InstanceMethods)
     extend plugin::ClassMethods if defined?(plugin::ClassMethods)
+    FederationKit::Services::ExtensionApplicator.call(plugin, self)
 
     if plugin.respond_to?(:configure, true)
       plugin.configure(self, *args, &block)
@@ -27,9 +28,11 @@ end
 
 require 'federation_kit/version'
 require 'federation_kit/error'
-require 'federation_kit/service'
 require 'federation_kit/plugin_errors'
 require 'federation_kit/plugin_registry'
+require 'federation_kit/service'
+require 'federation_kit/services/string_classifier'
+require 'federation_kit/services/extension_applicator'
 require 'federation_kit/plugins/base'
 
 FederationKit.plugin FederationKit::Plugins::Base