+++ spec/federation_kit/services/extension_applicator_spec.rb
@@ -1,0 +1,15 @@
+RSpec.describe FederationKit::Services::ExtensionApplicator do
+  describe '#call' do
+    let(:plugin) do
+      module TestPlugin
+        def self.extends; [FederationKit::Object, :actor]; end
+        module ObjectInstanceMethods; end
+        module ObjectClassMethods; end
+        module ActorInstanceMethods; end
+      end
+
+      TestPlugin
+    end
+
+  end
+end